# internal/models/podcast.go
package models

import (
	"time"

	"ikoyhn/podcast-sponsorblock/internal/services/common"
	"google.golang.org/api/youtube/v3"
)

type PodcastEpisode struct {
	Id                 int32         `gorm:"autoIncrement;primary_key;not null"`
	YoutubeVideoId     string        `json:"youtube_video_id" gorm:"index:youtubevideoid_type"`
	EpisodeName        string        `json:"episode_name"`
	EpisodeDescription string        `json:"episode_description"`
	PublishedDate      time.Time     `json:"published_date" gorm:"index"`
	Type               string        `json:"type" gorm:"index:youtubevideoid_type_channelid_type"`
	PodcastId          string        `json:"podcast_id" gorm:"index;foreignkey:PodcastId;association_foreignkey:Id"`
	Duration           time.Duration `json:"duration"`
}

type Podcast struct {
	Id              string           `json:"id" gorm:"primary_key"`
	AppleId         string           `json:"apple_id"`
	PodcastName     string           `json:"podcast_name"`
	Description     string           `json:"description"`
	Category        string           `json:"category"`
	PostedDate      string           `json:"posted_date"`
	ImageUrl        string           `json:"image_url"`
	LastBuildDate   string           `json:"last_build_date"`
	PodcastEpisodes []PodcastEpisode `json:"podcast_episodes"`
	ArtistName      string           `json:"artist_name"`
	Explicit        string           `json:"explicit"`
}

type EpisodePlaybackHistory struct {
	YoutubeVideoId   string  `json:"youtube_video_id" gorm:"primary_key"`
	LastAccessDate   int64   `json:"last_access_date" gorm:"index"`
	TotalTimeSkipped float64 `json:"total_time_skipped"`
}

type Analytics struct {
	Id              int32     `gorm:"autoIncrement;primary_key;not null"`
	EpisodeId       string    `json:"episode_id" gorm:"index;not null"`
	PlayCount       int64     `json:"play_count" gorm:"default:0"`
	TotalListenTime float64   `json:"total_listen_time" gorm:"default:0"` // in seconds
	LastPlayed      time.Time `json:"last_played" gorm:"index"`
	IPAddress       string    `json:"ip_address" gorm:"index"`
	Country         string    `json:"country" gorm:"index"`
	CreatedAt       time.Time `json:"created_at" gorm:"autoCreateTime"`
	UpdatedAt       time.Time `json:"updated_at" gorm:"autoUpdateTime"`
}

type Transcript struct {
	Id             int32     `gorm:"autoIncrement;primary_key;not null"`
	EpisodeId      string    `json:"episode_id" gorm:"uniqueIndex:idx_episode_lang;not null"`
	Language       string    `json:"language" gorm:"uniqueIndex:idx_episode_lang;not null;default:'en'"`
	Content        string    `json:"content" gorm:"type:text"`
	IsAutoGenerated bool     `json:"is_auto_generated" gorm:"default:false"`
	FetchedAt      time.Time `json:"fetched_at" gorm:"autoCreateTime"`
	UpdatedAt      time.Time `json:"updated_at" gorm:"autoUpdateTime"`
}

func NewPodcastEpisodeFromPlaylist(youtubeVideo *youtube.PlaylistItem) PodcastEpisode {
	// For PlaylistItems, use VideoPublishedAt for the actual publication date
	publishedAt := common.ParseYouTubeDate(youtubeVideo.ContentDetails.VideoPublishedAt)

	return PodcastEpisode{
		YoutubeVideoId:     youtubeVideo.Snippet.ResourceId.VideoId,
		EpisodeName:        youtubeVideo.Snippet.Title,
		EpisodeDescription: youtubeVideo.Snippet.Description,
		PublishedDate:      publishedAt,
		Type:               "PLAYLIST",
		PodcastId:          youtubeVideo.Snippet.PlaylistId,
	}
}

func NewPodcastEpisodeFromSearch(youtubeVideo *youtube.Video, duration time.Duration) PodcastEpisode {
	publishedAt := common.ParseYouTubeDate(youtubeVideo.Snippet.PublishedAt)

	return PodcastEpisode{
		YoutubeVideoId:     youtubeVideo.Id,
		EpisodeName:        youtubeVideo.Snippet.Title,
		EpisodeDescription: youtubeVideo.Snippet.Description,
		PublishedDate:      publishedAt,
		Type:               "CHANNEL",
		PodcastId:          youtubeVideo.Snippet.ChannelId,
		Duration:           duration,
	}
}